Removed ProcessStorage and moved its CacheItem class into ExpressiveValidator.
