+++ src/UoN.ExpressiveAnnotations.NetCore/Validators/ExpressiveValidator.cs
@@ -14,7 +14,6 @@ using Microsoft.AspNetCore.Mvc.ModelBinding;
 using Microsoft.AspNetCore.Mvc.ModelBinding.Validation;
 using Microsoft.Extensions.Caching.Memory;
 using UoN.ExpressiveAnnotations.NetCore.Attributes;
-using UoN.ExpressiveAnnotations.NetCore.Caching;
 
 namespace UoN.ExpressiveAnnotations.NetCore.Validators
 {
@@ -309,4 +308,14 @@ namespace UoN.ExpressiveAnnotations.NetCore.Validators
             }
         }
     }
+
+    internal class CacheItem
+    {
+        public IDictionary<string, string> FieldsMap { get; set; }
+        public IDictionary<string, object> ConstsMap { get; set; }
+        public IDictionary<string, object> EnumsMap { get; set; }
+        public IEnumerable<string> MethodsList { get; set; }
+        public IDictionary<string, string> ParsersMap { get; set; }
+    }
+
 }
